pyproject.toml
[project]
name = "reactshearchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.15",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "openai>=2.22.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
